backend/requirements.txt
django
djangorestframework
django-cors-headers
drf-extensions
django-redis